+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # stories
 
+## 0.1.4
+
+### Patch Changes
+
+- @shallot-ui/next@4.0.3
+
 ## 0.1.3
 
 ### Patch Changes