+++ components/Button/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shallot-ui/component-button
 
+## 3.0.2
+
+### Patch Changes
+
+- Improves ref forwarding
+
 ## 3.0.1
 
 ### Patch Changes