+++ packages/web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @shallot-ui/web
 
+## 4.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/component-button@3.0.2
+  - @shallot-ui/platform-react@2.0.3
+  - @shallot-ui/core-props@2.0.3
+
 ## 4.0.2
 
 ### Patch Changes